cleans up calendar files
